+++ User_application/ACMSim.h
@@ -63,6 +63,7 @@
 #include "pmsm_comm.h"
 #include "Bezier.h"
 #include "regulator_speedInnerLoop.h"
+#include "inverter_Compensation.h"
 
 // Header for global_variabels_definition.c
 struct GlobalWatch
@@ -86,7 +87,7 @@ double fabs(double x);
 REAL _lpf(REAL x, REAL y_tminus1, REAL time_const_inv);
 REAL PostionSpeedMeasurement_MovingAvergage(int32 QPOSCNT, st_enc *p_enc);
 extern REAL one_over_six;
-double difference_between_two_angles(double first, double second);
+REAL difference_between_two_angles(REAL first, REAL second);
